+++ backend/collaboration/src/config.ts
@@ -4,4 +4,10 @@ export const UI_HOST = process.env.PEERPREP_UI_HOST!;
 
 export const EXPRESS_PORT = process.env.EXPRESS_PORT;
 
-export const WEBSOCKET_PORT = process.env.WEBSOCKET_PORT;
+export const dbConfig = {
+  host: process.env.EXPRESS_DB_HOST!,
+  port: Number.parseInt(process.env.EXPRESS_DB_PORT!),
+  database: process.env.POSTGRES_DB!,
+  user: process.env.POSTGRES_USER,
+  password: process.env.POSTGRES_PASSWORD,
+};
