+++ backend/collaboration/src/server.ts
@@ -10,7 +10,13 @@ import { config, db } from '@/lib/db';
 import { logger } from '@/lib/utils';
 import roomRoutes from '@/routes/room';
 
+import { WebSocketServer } from 'ws';
+import { createServer } from 'http';
+import { setupWSConnection } from './y-postgresql-util/utils';
+import { setUpPersistence } from './y-postgresql-util/persistence';
 const app = express();
+const server = createServer(app);
+
 app.use(pino());
 app.use(json());
 app.use(
@@ -25,6 +31,10 @@ app.use('/room', roomRoutes);
 // Health Check for Docker
 app.get('/health', (_req, res) => res.status(StatusCodes.OK).send('OK'));
 
+// y-websocket server
+const wss = new WebSocketServer({ server });
+wss.on('connection', setupWSConnection);
+setUpPersistence();
 export const dbHealthCheck = async () => {
   try {
     await db`SELECT 1`;
@@ -42,4 +52,5 @@ app.get('/test-db', async (_req, res) => {
   await dbHealthCheck();
   res.json({ message: 'OK ' });
 });
-export default app;
+
+export default server;
